% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifify-package.R
\docType{package}
\name{gifify-package}
\alias{gifify}
\alias{gifify-package}
\title{gifify: Take Video Files and Turn Them into GIFs}
\description{
Take video files and turn them into GIFs with subtitles
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/imaddowzimet/gifify}
  \item Report bugs at \url{https://github.com/imaddowzimet/gifify/issues}
}

}
\author{
\strong{Maintainer}: Isaac Maddow-Zimet \email{imaddowzimet@gmail.com} (YOUR-ORCID-ID)

}
\keyword{internal}
