% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifify.R
\name{gifify}
\alias{gifify}
\title{Convert video to GIF (with subtitles)}
\usage{
gifify(file, text = "", cut = NULL, gifname = "brilliant.gif",
  width = "400", color = "yellow", size = 22,
  strokecolor = "black", gravity = "south", style = "italic",
  weight = 700, delay = 1/20)
}
\arguments{
\item{file}{File path to video file}

\item{text}{Subtitle text (one or two text strings)}

\item{cut}{If using multiple subtitles, the frame of the video when the subtitles should switch (might take some experimentation)}

\item{gifname}{Name of exported file. Defaults to "brilliant.gif"}

\item{width}{width of exported file}

\item{color}{color of subtitle text}

\item{size}{size of subtitle text}

\item{gravity}{position of subtitle text. "south" is default.}

\item{style}{font family.}

\item{weight}{thickness of font.}

\item{delay}{sets speed of GIF; default (1/20) should be normal speed.}
}
\value{
Returns a brilliant GIF.
}
\description{
`gifify` lets you import video files and create subtitled GIFs from them, for all of your meme-ing needs.
}
